+++ src/core/lombok/javac/handlers/JavacHandlerUtil.java
@@ -42,17 +42,18 @@ import lombok.core.AnnotationValues.AnnotationValue;
 import lombok.core.TransformationsUtil;
 import lombok.core.TypeResolver;
 import lombok.experimental.Accessors;
-import lombok.javac.Javac;
 import lombok.javac.JavacNode;
 
+import com.sun.tools.javac.code.BoundKind;
 import com.sun.tools.javac.code.Flags;
-import com.sun.tools.javac.code.TypeTags;
 import com.sun.tools.javac.tree.JCTree;
 import com.sun.tools.javac.tree.JCTree.JCAnnotation;
+import com.sun.tools.javac.tree.JCTree.JCArrayTypeTree;
 import com.sun.tools.javac.tree.JCTree.JCAssign;
 import com.sun.tools.javac.tree.JCTree.JCClassDecl;
 import com.sun.tools.javac.tree.JCTree.JCCompilationUnit;
 import com.sun.tools.javac.tree.JCTree.JCExpression;
+import com.sun.tools.javac.tree.JCTree.JCFieldAccess;
 import com.sun.tools.javac.tree.JCTree.JCIdent;
 import com.sun.tools.javac.tree.JCTree.JCImport;
 import com.sun.tools.javac.tree.JCTree.JCLiteral;
@@ -60,14 +61,20 @@ import com.sun.tools.javac.tree.JCTree.JCMethodDecl;
 import com.sun.tools.javac.tree.JCTree.JCMethodInvocation;
 import com.sun.tools.javac.tree.JCTree.JCModifiers;
 import com.sun.tools.javac.tree.JCTree.JCNewArray;
+import com.sun.tools.javac.tree.JCTree.JCPrimitiveTypeTree;
 import com.sun.tools.javac.tree.JCTree.JCStatement;
+import com.sun.tools.javac.tree.JCTree.JCTypeParameter;
+import com.sun.tools.javac.tree.JCTree.JCTypeApply;
 import com.sun.tools.javac.tree.JCTree.JCVariableDecl;
+import com.sun.tools.javac.tree.JCTree.JCWildcard;
+import com.sun.tools.javac.tree.JCTree.TypeBoundKind;
 import com.sun.tools.javac.tree.TreeMaker;
 import com.sun.tools.javac.tree.TreeScanner;
 import com.sun.tools.javac.util.JCDiagnostic.DiagnosticPosition;
 import com.sun.tools.javac.util.List;
 import com.sun.tools.javac.util.ListBuffer;
 import com.sun.tools.javac.util.Name;
+import com.sun.tools.javac.util.Options;
 
 /**
  * Container for static utility methods useful to handlers written for javac.
@@ -91,6 +98,16 @@ public class JavacHandlerUtil {
 	}
 	
 	private static Map<JCTree, WeakReference<JCTree>> generatedNodes = new WeakHashMap<JCTree, WeakReference<JCTree>>();
+	
+	/**
+	 * Contributed by Jan Lahoda; many lombok transformations should not be run (or a lite version should be run) when the netbeans editor
+	 * is running javac on the open source file to find inline errors and such. As class files are compiled separately this does not affect
+	 * actual runtime behaviour or file output of the netbeans IDE.
+	 */
+	public static boolean inNetbeansEditor(JavacNode node) {
+		Options options = Options.instance(node.getContext());
+		return (options.keySet().contains("ide") && !options.keySet().contains("backgroundCompilation"));
+	}
 	
 	public static JCTree getGeneratedBy(JCTree node) {
 		synchronized (generatedNodes) {
@@ -116,6 +133,35 @@ public class JavacHandlerUtil {
 			else generatedNodes.put(node, new WeakReference<JCTree>(source));
 		}
 		return node;
+	}
+	
+	public static boolean hasAnnotation(Class<? extends Annotation> type, JavacNode node) {
+		return hasAnnotation(type, node, false);
+	}
+	
+	public static boolean hasAnnotationAndDeleteIfNeccessary(Class<? extends Annotation> type, JavacNode node) {
+		return hasAnnotation(type, node, true);
+	}
+	
+	private static boolean hasAnnotation(Class<? extends Annotation> type, JavacNode node, boolean delete) {
+		if (node == null) return false;
+		if (type == null) return false;
+		switch (node.getKind()) {
+		case ARGUMENT:
+		case FIELD:
+		case LOCAL:
+		case TYPE:
+		case METHOD:
+			for (JavacNode child : node.down()) {
+				if (annotationTypeMatches(type, child)) {
+					if (delete) deleteAnnotationIfNeccessary(child, type);
+					return true;
+				}
+			}
+			// intentional fallthrough
+		default:
+			return false;
+		}
 	}
 	
 	/**
@@ -232,6 +278,7 @@ public class JavacHandlerUtil {
 	 * Only does this if the DeleteLombokAnnotations class is in the context.
 	 */
 	public static void deleteAnnotationIfNeccessary(JavacNode annotation, Class<? extends Annotation> annotationType) {
+		if (inNetbeansEditor(annotation)) return;
 		if (!annotation.shouldDeleteLombokAnnotations()) return;
 		JavacNode parentNode = annotation.directUp();
 		switch (parentNode.getKind()) {
@@ -262,6 +309,7 @@ public class JavacHandlerUtil {
 	}
 	
 	public static void deleteImportFromCompilationUnit(JavacNode node, String name) {
+		if (inNetbeansEditor(node)) return;
 		if (!node.shouldDeleteLombokAnnotations()) return;
 		ListBuffer<JCTree> newDefs = ListBuffer.lb();
 		
@@ -296,11 +344,7 @@ public class JavacHandlerUtil {
 	 * Convenient wrapper around {@link TransformationsUtil#toAllGetterNames(lombok.core.AnnotationValues, CharSequence, boolean)}.
 	 */
 	public static java.util.List<String> toAllGetterNames(JavacNode field) {
-		String fieldName = field.getName();
-		boolean isBoolean = isBoolean(field);
-		AnnotationValues<Accessors> accessors = JavacHandlerUtil.getAccessorsForField(field);
-		
-		return TransformationsUtil.toAllGetterNames(accessors, fieldName, isBoolean);
+		return TransformationsUtil.toAllGetterNames(getAccessorsForField(field), field.getName(), isBoolean(field));
 	}
 	
 	/**
@@ -309,11 +353,7 @@ public class JavacHandlerUtil {
 	 * Convenient wrapper around {@link TransformationsUtil#toGetterName(lombok.core.AnnotationValues, CharSequence, boolean)}.
 	 */
 	public static String toGetterName(JavacNode field) {
-		String fieldName = field.getName();
-		boolean isBoolean = isBoolean(field);
-		AnnotationValues<Accessors> accessors = JavacHandlerUtil.getAccessorsForField(field);
-		
-		return TransformationsUtil.toGetterName(accessors, fieldName, isBoolean);
+		return TransformationsUtil.toGetterName(getAccessorsForField(field), field.getName(), isBoolean(field));
 	}
 	
 	/**
@@ -321,11 +361,7 @@ public class JavacHandlerUtil {
 	 * Convenient wrapper around {@link TransformationsUtil#toAllSetterNames(lombok.core.AnnotationValues, CharSequence, boolean)}.
 	 */
 	public static java.util.List<String> toAllSetterNames(JavacNode field) {
-		String fieldName = field.getName();
-		boolean isBoolean = isBoolean(field);
-		AnnotationValues<Accessors> accessors = JavacHandlerUtil.getAccessorsForField(field);
-		
-		return TransformationsUtil.toAllSetterNames(accessors, fieldName, isBoolean);
+		return TransformationsUtil.toAllSetterNames(getAccessorsForField(field), field.getName(), isBoolean(field));
 	}
 	
 	/**
@@ -334,11 +370,24 @@ public class JavacHandlerUtil {
 	 * Convenient wrapper around {@link TransformationsUtil#toSetterName(lombok.core.AnnotationValues, CharSequence, boolean)}.
 	 */
 	public static String toSetterName(JavacNode field) {
-		String fieldName = field.getName();
-		boolean isBoolean = isBoolean(field);
-		AnnotationValues<Accessors> accessors = JavacHandlerUtil.getAccessorsForField(field);
-		
-		return TransformationsUtil.toSetterName(accessors, fieldName, isBoolean);
+		return TransformationsUtil.toSetterName(getAccessorsForField(field), field.getName(), isBoolean(field));
+	}
+	
+	/**
+	 * Translates the given field into all possible wither names.
+	 * Convenient wrapper around {@link TransformationsUtil#toAllWitherNames(lombok.core.AnnotationValues, CharSequence, boolean)}.
+	 */
+	public static java.util.List<String> toAllWitherNames(JavacNode field) {
+		return TransformationsUtil.toAllWitherNames(getAccessorsForField(field), field.getName(), isBoolean(field));
+	}
+	
+	/**
+	 * @return the likely wither name for the stated field. (e.g. private boolean foo; to withFoo).
+	 * 
+	 * Convenient wrapper around {@link TransformationsUtil#toWitherName(lombok.core.AnnotationValues, CharSequence, boolean)}.
+	 */
+	public static String toWitherName(JavacNode field) {
+		return TransformationsUtil.toWitherName(getAccessorsForField(field), field.getName(), isBoolean(field));
 	}
 	
 	/**
@@ -353,6 +402,26 @@ public class JavacHandlerUtil {
 		boolean forced = (accessors.getActualExpression("chain") != null);
 		Accessors instance = accessors.getInstance();
 		return instance.chain() || (instance.fluent() && !forced);
+	}
+	
+	public static JCExpression cloneSelfType(JavacNode field) {
+		JavacNode typeNode = field;
+		TreeMaker maker = field.getTreeMaker();
+		while (typeNode != null && typeNode.getKind() != Kind.TYPE) typeNode = typeNode.up();
+		if (typeNode != null && typeNode.get() instanceof JCClassDecl) {
+			JCClassDecl type = (JCClassDecl) typeNode.get();
+			ListBuffer<JCExpression> typeArgs = ListBuffer.lb();
+			if (!type.typarams.isEmpty()) {
+				for (JCTypeParameter tp : type.typarams) {
+					typeArgs.append(maker.Ident(tp.name));
+				}
+				return maker.TypeApply(maker.Ident(type.name), typeArgs.toList());
+			} else {
+				return maker.Ident(type.name);
+			}
+		} else {
+			return null;
+		}
 	}
 	
 	private static boolean isBoolean(JavacNode field) {
@@ -555,7 +624,7 @@ public class JavacHandlerUtil {
 		return null;
 	}
 	
-	enum FieldAccess {
+	public enum FieldAccess {
 		GETTER, PREFER_FIELD, ALWAYS_FIELD;
 	}
 	
@@ -735,6 +804,8 @@ public class JavacHandlerUtil {
 	 * In javac, dotted access of any kind, from {@code java.lang.String} to {@code var.methodName}
 	 * is represented by a fold-left of {@code Select} nodes with the leftmost string represented by
 	 * a {@code Ident} node. This method generates such an expression.
+	 * <p>
+	 * The position of the generated node(s) will be unpositioned (-1).
 	 * 
 	 * For example, maker.Select(maker.Select(maker.Ident(NAME[java]), NAME[lang]), NAME[String]).
 	 * 
@@ -742,12 +813,30 @@ public class JavacHandlerUtil {
 	 * @see com.sun.tools.javac.tree.JCTree.JCFieldAccess
 	 */
 	public static JCExpression chainDots(JavacNode node, String... elems) {
+		return chainDots(node, -1, elems);
+	}
+	
+	/**
+	 * In javac, dotted access of any kind, from {@code java.lang.String} to {@code var.methodName}
+	 * is represented by a fold-left of {@code Select} nodes with the leftmost string represented by
+	 * a {@code Ident} node. This method generates such an expression.
+	 * <p>
+	 * The position of the generated node(s) will be equal to the {@code pos} parameter.
+	 *
+	 * For example, maker.Select(maker.Select(maker.Ident(NAME[java]), NAME[lang]), NAME[String]).
+	 * 
+	 * @see com.sun.tools.javac.tree.JCTree.JCIdent
+	 * @see com.sun.tools.javac.tree.JCTree.JCFieldAccess
+	 */
+	public static JCExpression chainDots(JavacNode node, int pos, String... elems) {
 		assert elems != null;
 		assert elems.length > 0;
 		
-		JCExpression e = node.getTreeMaker().Ident(node.toName(elems[0]));
+		TreeMaker maker = node.getTreeMaker();
+		if (pos != -1) maker = maker.at(pos);
+		JCExpression e = maker.Ident(node.toName(elems[0]));
 		for (int i = 1 ; i < elems.length ; i++) {
-			e = node.getTreeMaker().Select(e, node.toName(elems[i]));
+			e = maker.Select(e, node.toName(elems[i]));
 		}
 		
 		return e;
@@ -765,7 +854,7 @@ public class JavacHandlerUtil {
 	 * @see com.sun.tools.javac.tree.JCTree.JCFieldAccess
 	 */
 	public static JCExpression chainDotsString(JavacNode node, String elems) {
-		return chainDots(node, elems.split("\\."));	
+		return chainDots(node, elems.split("\\."));
 	}
 	
 	/**
@@ -800,7 +889,7 @@ public class JavacHandlerUtil {
 		JCExpression npe = chainDots(variable, "java", "lang", "NullPointerException");
 		JCTree exception = treeMaker.NewClass(null, List.<JCExpression>nil(), npe, List.<JCExpression>of(treeMaker.Literal(fieldName.toString())), null);
 		JCStatement throwStatement = treeMaker.Throw(exception);
-		return treeMaker.If(treeMaker.Binary(Javac.getCtcInt(JCTree.class, "EQ"), treeMaker.Ident(fieldName), treeMaker.Literal(Javac.getCtcInt(TypeTags.class, "BOT"), null)), throwStatement, null);
+		return treeMaker.If(treeMaker.Binary(CTC_EQUAL, treeMaker.Ident(fieldName), treeMaker.Literal(CTC_BOT, null)), throwStatement, null);
 	}
 	
 	/**
@@ -888,5 +977,72 @@ public class JavacHandlerUtil {
 		while ((node != null) && !(node.get() instanceof JCClassDecl)) node = node.up();
 		
 		return node;
+	}
+	
+	/**
+	 * Creates a full clone of a given javac AST type node. Every part is cloned (every identifier, every select, every wildcard, every type apply).
+	 * 
+	 * If there's any node in the tree that we don't know how to clone, that part isn't cloned. However, we wouldn't know what could possibly show up that we
+	 * can't currently clone; that's just a safeguard.
+	 * 
+	 * This should be used if the type looks the same in the code, but resolves differently. For example, a static method that has some generics in it named after
+	 * the class's own parameter, but as its a static method, the static method's notion of {@code T} is different from the class notion of {@code T}. If you're duplicating
+	 * a type used in the class context, you need to use this method.
+	 */
+	public static JCExpression cloneType(TreeMaker maker, JCExpression in, JCTree source) {
+		JCExpression out = cloneType0(maker, in);
+		if (out != null) recursiveSetGeneratedBy(out, source);
+		return out;
+	}
+	
+	private static JCExpression cloneType0(TreeMaker maker, JCTree in) {
+		if (in == null) return null;
+		
+		if (in instanceof JCPrimitiveTypeTree) return (JCExpression) in;
+		
+		if (in instanceof JCIdent) {
+			return maker.Ident(((JCIdent) in).name);
+		}
+		
+		if (in instanceof JCFieldAccess) {
+			JCFieldAccess fa = (JCFieldAccess) in;
+			return maker.Select(cloneType0(maker, fa.selected), fa.name);
+		}
+		
+		if (in instanceof JCArrayTypeTree) {
+			JCArrayTypeTree att = (JCArrayTypeTree) in;
+			return maker.TypeArray(cloneType0(maker, att.elemtype));
+		}
+		
+		if (in instanceof JCTypeApply) {
+			JCTypeApply ta = (JCTypeApply) in;
+			ListBuffer<JCExpression> lb = ListBuffer.lb();
+			for (JCExpression typeArg : ta.arguments) {
+				lb.append(cloneType0(maker, typeArg));
+			}
+			return maker.TypeApply(cloneType0(maker, ta.clazz), lb.toList());
+		}
+		
+		if (in instanceof JCWildcard) {
+			JCWildcard w = (JCWildcard) in;
+			JCExpression newInner = cloneType0(maker, w.inner);
+			TypeBoundKind newKind;
+			switch (w.getKind()) {
+			case SUPER_WILDCARD:
+				newKind = maker.TypeBoundKind(BoundKind.SUPER);
+				break;
+			case EXTENDS_WILDCARD:
+				newKind = maker.TypeBoundKind(BoundKind.EXTENDS);
+				break;
+			default:
+			case UNBOUNDED_WILDCARD:
+				newKind = maker.TypeBoundKind(BoundKind.UNBOUND);
+				break;
+			}
+			return maker.Wildcard(newKind, newInner);
+		}
+		
+		// This is somewhat unsafe, but it's better than outright throwing an exception here. Returning null will just cause an exception down the pipeline.
+		return (JCExpression) in;
 	}
 }